Create Procesamiento de imágenes
A code designed for image processing with specific application in the categorization of nanomaterials is included. This code performs several modifications, such as the conversion of images from color scale (RGB) to gray scale, and the automation in the creation of folders to store images classified in different categories according to their processing.